refactor(product): settings kit descent — kit members move to primitives/patterns/settings
Mechanical move only: no visual changes, no new components, no prop changes.
The settings kit named in specs/DESIGN_SYSTEM.md now owns a subdirectory under
primitives/patterns/, mirroring the composer/toast/sidebar move in #1781.

  patterns/settings/  SettingsGroup, SettingsMenu   (already in this tier)
                      SettingsSection, SettingsRow, SettingsScopeTabs,
                      SettingsSaveFooter, SettingsEmptyState
                      (+ the colocated SettingsSection.test.tsx)

All moves are git renames. The only content edits inside a moved file are the
three intra-kit import specifiers in SettingsSection.tsx and its test.
SettingsGroup.tsx moves byte-for-byte: its fragment/null-child divider handling
from the settings wash restyle (#1777) is untouched.

Tier migration: the five patterns above came from components/patterns/ (the
domain-aware tier) and none of them import domain code, so they descend to the
pattern tier per the kit-cohesion rule. Their registry entries and demo bodies
move verbatim from product-patterns.tsx to patterns.tsx; pinned counts go
36/24/10 (70) -> 36/29/10 (75).

SettingsPageHeader is deliberately NOT moved: the settings slice retires it in
favour of PageHeader variant="flat", so moving it would be churn ahead of a
delete. It stays in components/patterns/ with its registry entry and index row
in place, and Current Gaps records it as the remaining settings-kit split.

Consumers updated (51 files) to #product/primitives/patterns/settings/<Name>.

Docs: sanctioned-index rows repointed and regrouped into the pattern-tier
table, the area-kit paragraph names the settings members, the domain-aware
tier paragraph drops the retired settings-family sentence, and the Current
Gaps kit bullet is reworded to post-move truth. The settings
information-architecture doc's five path references follow.

Gates run, all green:
  python3 scripts/check_docs.py
  python3 scripts/check_frontend_boundaries.py
  python3 scripts/check_appearance_scaling.py
  python3 scripts/report_frontend_structure.py --strict --summary-only
  vitest: library-registry (4) + SettingsSection (8) + SettingsGroup (5)
          + 5 consumer suites; 38 tests, 7 files green

Baselines: appearance_scaling_baseline.json, frontend_structure_allowlist.txt
and frontend_boundaries_allowlist.txt carry zero entries keyed by the moved
files, so no keys were renamed (verified, not assumed).
