+++ apps/packages/product-client/src/components/settings/panes/BillingPane.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import { BillingSettingsSurface } from "#product/components/settings/panes/billing/BillingSettingsSurface";
 import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
-import { SettingsEmptyState } from "#product/components/patterns/SettingsEmptyState";
+import { SettingsEmptyState } from "#product/primitives/patterns/settings/SettingsEmptyState";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
 
 import { PROLIFERATE_PRICING_URL } from "#product/config/capabilities";
