+++ apps/packages/product-client/src/components/settings/panes/GeneralPane.tsx
@@ -1,10 +1,10 @@
 import { useMemo } from "react";
 import { useShallow } from "zustand/react/shallow";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
-import { SETTINGS_CONTROL_WIDTH_CLASS, SettingsRow } from "#product/components/patterns/SettingsRow";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
+import { SETTINGS_CONTROL_WIDTH_CLASS, SettingsRow } from "#product/primitives/patterns/settings/SettingsRow";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
 import { Button } from "#product/primitives/Button";
-import { SettingsMenu } from "#product/primitives/patterns/SettingsMenu";
+import { SettingsMenu } from "#product/primitives/patterns/settings/SettingsMenu";
 import { Switch } from "#product/primitives/Switch";
 import { OpenTargetIcon } from "#product/components/workspace/open-target/OpenTargetIcon";
 import { useAvailableEditors } from "#product/hooks/access/tauri/shell/use-available-editors";
