+++ apps/packages/product-client/src/components/settings/panes/OrganizationIntegrationsPane.tsx
@@ -3,10 +3,10 @@ import { Badge } from "#product/primitives/Badge";
 import { Button } from "#product/primitives/Button";
 import { Input } from "#product/primitives/Input";
 import { Switch } from "#product/primitives/Switch";
-import { SettingsEmptyState } from "#product/components/patterns/SettingsEmptyState";
+import { SettingsEmptyState } from "#product/primitives/patterns/settings/SettingsEmptyState";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
-import { SettingsRow } from "#product/components/patterns/SettingsRow";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
+import { SettingsRow } from "#product/primitives/patterns/settings/SettingsRow";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
 import type { AdminIntegrationDefinition } from "@proliferate/cloud-sdk/client/integrations";
 import { AddCustomIntegrationDialog } from "#product/components/settings/panes/integrations/AddCustomIntegrationDialog";
 import { IntegrationIcon } from "#product/components/settings/panes/integrations/IntegrationIcon";
