+++ apps/packages/product-client/src/components/settings/panes/OrganizationMembersPane.tsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 import { Button } from "#product/primitives/Button";
 import { OrganizationInvitationsSection } from "#product/components/settings/panes/organization/OrganizationInvitationsSection";
 import { OrganizationMembersSection } from "#product/components/settings/panes/organization/OrganizationMembersSection";
-import { SettingsEmptyState } from "#product/components/patterns/SettingsEmptyState";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
+import { SettingsEmptyState } from "#product/primitives/patterns/settings/SettingsEmptyState";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
 import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
 import { useIsAdmin } from "#product/hooks/access/cloud/organizations/use-is-admin";
