+++ apps/packages/product-client/src/components/settings/panes/UserIntegrationsPane.tsx
@@ -3,9 +3,9 @@ import { useNavigate } from "react-router-dom";
 import { Button } from "#product/primitives/Button";
 import { ConfirmationDialog } from "#product/primitives/patterns/ConfirmationDialog";
 import { Input } from "#product/primitives/Input";
-import { SettingsEmptyState } from "#product/components/patterns/SettingsEmptyState";
+import { SettingsEmptyState } from "#product/primitives/patterns/settings/SettingsEmptyState";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
 import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
 import { IntegrationConnectDialog, type IntegrationConnectSubmit } from "#product/components/settings/panes/integrations/IntegrationConnectDialog";
 import { IntegrationRow } from "#product/components/settings/panes/integrations/IntegrationRow";
