+++ apps/packages/product-client/src/components/settings/panes/agents/harness/HarnessSettingsSection.tsx
@@ -1,7 +1,7 @@
 import { useCallback, useMemo } from "react";
 import type { AgentAuthSurface } from "@proliferate/cloud-sdk";
 import { Switch } from "#product/primitives/Switch";
-import { SettingsRow } from "#product/components/patterns/SettingsRow";
+import { SettingsRow } from "#product/primitives/patterns/settings/SettingsRow";
 import {
   usePutAuthSelections,
   useAgentAuthState,
