+++ apps/packages/product-client/src/components/settings/panes/environments/WorktreeStorageSection.tsx
@@ -3,8 +3,8 @@ import { Button } from "#product/primitives/Button";
 import { Input } from "#product/primitives/Input";
 import { Label } from "#product/primitives/Label";
 import { Minus, Plus } from "#product/primitives/icons/core";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
-import { SETTINGS_CONTROL_WIDTH_CLASS, SettingsRow } from "#product/components/patterns/SettingsRow";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
+import { SETTINGS_CONTROL_WIDTH_CLASS, SettingsRow } from "#product/primitives/patterns/settings/SettingsRow";
 import { RuntimePressureDetailsDialog } from "#product/components/workspace/chat/input/RuntimePressureDetailsDialog";
 import { RuntimePressureRing } from "#product/components/workspace/chat/input/RuntimePressureIndicator";
 import { useWorktreeCleanupPolicy } from "#product/hooks/workspaces/facade/use-worktree-cleanup-policy";
