+++ apps/packages/product-client/src/components/settings/panes/organization/OrganizationBillingLinkSection.tsx
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import { Button } from "#product/primitives/Button";
 import { ChevronRight } from "#product/primitives/icons/core";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
 import { buildSettingsHref } from "#product/lib/domain/settings/navigation";
 
 export function OrganizationBillingLinkSection() {
