+++ apps/packages/product-client/src/components/settings/panes/organization/OrganizationMembersSection.tsx
@@ -1,7 +1,7 @@
 import { useState } from "react";
 import { Search } from "#product/primitives/icons/core";
 import { Input } from "#product/primitives/Input";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
 import { OrganizationMembersList } from "#product/components/settings/panes/organization/OrganizationMembersList";
 import { OrganizationSelectMenu } from "#product/components/settings/panes/organization/OrganizationSelectMenu";
 import { buildMemberRows } from "#product/lib/domain/organizations/member-list-rows";
