+++ apps/packages/product-client/src/components/settings/panes/repo/RepoActionsPane.tsx
@@ -3,9 +3,9 @@ import type { SetupHint } from "@anyharness/sdk";
 import { useDetectRepoRootSetupQuery } from "@anyharness/sdk-react";
 import { ScriptBlock } from "#product/components/settings/panes/repo/ScriptBlock";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
-import { SettingsRow } from "#product/components/patterns/SettingsRow";
-import { SettingsSaveFooter } from "#product/components/patterns/SettingsSaveFooter";
-import { SettingsSection } from "#product/components/patterns/SettingsSection";
+import { SettingsRow } from "#product/primitives/patterns/settings/SettingsRow";
+import { SettingsSaveFooter } from "#product/primitives/patterns/settings/SettingsSaveFooter";
+import { SettingsSection } from "#product/primitives/patterns/settings/SettingsSection";
 import { Checkbox } from "#product/primitives/Checkbox";
 import { Input } from "#product/primitives/Input";
 import { Label } from "#product/primitives/Label";
