+++ apps/packages/product-client/src/components/settings/panes/repo/RepoEnvironmentPane.tsx
@@ -1,6 +1,6 @@
 import { KeyRound } from "lucide-react";
 import { SecretManagementPanel } from "#product/components/patterns/secrets/SecretManagementPanel";
-import { SettingsEmptyState } from "#product/components/patterns/SettingsEmptyState";
+import { SettingsEmptyState } from "#product/primitives/patterns/settings/SettingsEmptyState";
 import { SettingsPageHeader } from "#product/components/patterns/SettingsPageHeader";
 import { Button } from "#product/primitives/Button";
 import { useCloudSecretsPanel } from "#product/hooks/access/cloud/use-cloud-secrets-panel";
