+++ apps/packages/product-client/src/components/settings/screen/SettingsScreen.tsx
@@ -28,7 +28,7 @@ import {
 import { RepoScopeHeaderControls } from "#product/components/settings/screen/RepoScopeHeaderControls";
 import { AgentScopeHeaderControls } from "#product/components/settings/screen/AgentScopeHeaderControls";
 import { SettingsSidebar } from "#product/components/settings/sidebar/SettingsSidebar";
-import { SettingsScopeTabs } from "#product/components/patterns/SettingsScopeTabs";
+import { SettingsScopeTabs } from "#product/primitives/patterns/settings/SettingsScopeTabs";
 import { ArrowLeft } from "lucide-react";
 import { SETTINGS_COPY } from "#product/copy/settings/settings-copy";
 import { useCloudAvailabilityState } from "#product/hooks/cloud/derived/use-cloud-availability-state";
